+++ src/ODR/AdminBundle/Command/CryptoCommand.php
@@ -2,15 +2,14 @@
 
 /**
 * Open Data Repository Data Publisher
-* ImportWorker Command
+* Crypto Command
 * (C) 2015 by Nathan Stone (nate.stone@opendatarepository.org)
 * (C) 2015 by Alex Pires (ajpires@email.arizona.edu)
 * Released under the GPLv2
 *
 * This Symfony console command takes beanstalk jobs from the
-* import_worker tube and passes the parameters to XMLController,
-* which will attempt to import the contents of an XML file into
-* a DataRecord entity in the database.
+* crypto_requests tube and passes the parameters to WorkerController
+* to either encrypt a file/image entity, or decrypt a file entity.
 *
 */
 
@@ -25,19 +24,16 @@ use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
-// dunno if needed
-use ODR\AdminBundle\Entity\DataRecord;
-use ODR\AdminBundle\Entity\DataType;
 
-class ImportWorkerCommand extends ContainerAwareCommand
+class CryptoCommand extends ContainerAwareCommand
 {
     protected function configure()
     {
         parent::configure();
 
         $this
-            ->setName('odr_import:worker')
-            ->setDescription('Attempts to import an XML file into the database...');
+            ->setName('odr_crypto:worker')
+            ->setDescription('Encrypts/Decrypts a File or Image object');
     }
 
     protected function execute(InputInterface $input, OutputInterface $output)
@@ -45,41 +41,35 @@ class ImportWorkerCommand extends ContainerAwareCommand
         // Only need to load these once...
         $container = $this->getContainer();
         $logger = $container->get('logger');
-        $router = $container->get('router');
         $pheanstalk = $container->get('pheanstalk');
 
-        // Run command until manually stopped
         while (true) {
+            // Run command until manually stopped
             $job = null;
             try {
                 // Wait for a job?
-//                $job = $pheanstalk->watch($memcached_prefix.'_import_datarecord')->ignore('default')->reserve();
-                $job = $pheanstalk->watch('import_datarecord')->ignore('default')->reserve();
+                $job = $pheanstalk->watch('crypto_requests')->ignore('default')->reserve();
 
                 // Get Job Data
-                $data = json_decode($job->getData());
+                $data = json_decode($job->getData()); 
 
                 // 
-                $str = '';
-                if ($data->datatype_id !== '')
-                    $str = 'ImportWorkerCommand.php: Attempting to import "'.$data->xml_filename.'" for DataType '.$data->datatype_id.' from '.$data->memcached_prefix.'...';
-                else if ($data->datafield_id !== '')
-                    $str = 'ImportWorkerCommand.php: Attempting to import "'.$data->xml_filename.'" for DataField '.$data->datafield_id.' from '.$data->memcached_prefix.'...';
-                else
-                    throw new \Exception('Invalid job data');
-
+                $logger->info('CryptoCommand.php: '.$data->crypto_type.' request for '.$data->object_type.' '.$data->object_id.' from '.$data->memcached_prefix.'...');
                 $current_time = new \DateTime();
-                $output->writeln( $current_time->format('Y-m-d H:i:s').' (UTC-5)' );
-                $output->writeln($str);
-                $logger->info('ImportWorkerCommand.php: '.$str);
+                $output->writeln( $current_time->format('Y-m-d H:i:s').' (UTC-5)' );                
+                $output->writeln($data->crypto_type.' request for '.$data->object_type.' '.$data->object_id.' from '.$data->memcached_prefix.'...');
 
                 // Need to use cURL to send a POST request...thanks symfony
                 $ch = curl_init();
 
-$output->writeln($data->url);
-
-                // Create the required url and the parameters to send
-                $parameters = array('datatype_id' => $data->datatype_id, 'datafield_id' => $data->datafield_id, 'api_key' => $data->api_key, 'user_id' => $data->user_id, 'xml_filename' => $data->xml_filename);
+                // Create the required parameters to send
+                $parameters = array(
+                    'object_type' => $data->object_type,
+                    'object_id' => $data->object_id,
+                    'target_filepath' => $data->target_filepath,
+                    'crypto_type' => $data->crypto_type,
+                    'api_key' => $data->api_key
+                );
 
                 // Set the options for the POST request
                 curl_setopt_array($ch, array(
@@ -89,14 +79,20 @@ $output->writeln($data->url);
                         CURLOPT_FRESH_CONNECT => 1,
                         CURLOPT_RETURNTRANSFER => 1,
                         CURLOPT_FORBID_REUSE => 1,
-                        CURLOPT_TIMEOUT => 300,
+                        CURLOPT_TIMEOUT => 0,   // TODO - actual timeout value instead of "never"?
                         CURLOPT_POSTFIELDS => http_build_query($parameters)
                     )
                 );
 
                 // Send the request
                 if( ! $ret = curl_exec($ch)) {
-                    throw new \Exception( curl_error($ch) );
+                    if (curl_errno($ch) == 6) {
+                        // Could not resolve host
+                        throw new \Exception('retry');
+                    }
+                    else {
+                        throw new \Exception( curl_error($ch) );
+                    }
                 }
 
                 // Do things with the response returned by the controller?
@@ -111,23 +107,38 @@ $output->writeln($data->url);
                     // Should always be a json return...
                     throw new \Exception( print_r($ret, true) );
                 }
-//$logger->debug('ImportWorkerCommand.php: curl results...'.print_r($result, true));
+//$logger->debug('EncryptCommand.php: curl results...'.print_r($result, true));
 
                 // Done with this cURL object
                 curl_close($ch);
 
-                // Dealt with (or ignored) the job
+                // Dealt with the job
                 $pheanstalk->delete($job);
+
+                // Sleep for a bit
+                usleep(200000);
 
             }
             catch (\Exception $e) {
-                $output->writeln($e->getMessage());
-                $logger->err('ImportWorkerCommand.php: '.$e->getMessage());
+                if ( $e->getMessage() == 'retry' ) {
+                    $output->writeln( 'Could not resolve host, releasing job to try again' );
+                    $logger->err('CryptoCommand.php: '.$e->getMessage());
 
-                // Delete the job so the queue doesn't hang, in theory
-                $pheanstalk->delete($job);
+                    // Release the job back into the ready queue to try again
+                    $pheanstalk->release($job);
+
+                    // Sleep for a bit
+                    usleep(1000000);     // sleep for 1 second
+                }
+                else {
+                    $output->writeln($e->getMessage());
+
+                    $logger->err('CryptoCommand.php: '.$e->getMessage());
+
+                    // Delete the job so the queue doesn't hang, in theory
+                    $pheanstalk->delete($job);
+                }
             }
         }
     }
-
 }
