+++ src/ODR/AdminBundle/Command/MigrateCommand.php
@@ -45,7 +45,6 @@ class MigrateCommand extends ContainerAwareCommand
     {
         // Only need to load these once...
         $container = $this->getContainer();
-        $router = $container->get('router');
         $logger = $container->get('logger');
         $pheanstalk = $container->get('pheanstalk');
 
@@ -54,8 +53,7 @@ class MigrateCommand extends ContainerAwareCommand
             $job = null;
             try {
                 // Wait for a job?
-//                $job = $pheanstalk->watch($memcached_prefix.'_migrate_datafields')->ignore('default')->reserve(); 
-                $job = $pheanstalk->watch('migrate_datafields')->ignore('default')->reserve(); 
+                $job = $pheanstalk->watch('migrate_datafields')->ignore('default')->reserve();
 
                 // Get Job Data
                 $data = json_decode($job->getData()); 
@@ -95,7 +93,13 @@ class MigrateCommand extends ContainerAwareCommand
 
                 // Send the request
                 if( ! $ret = curl_exec($ch)) {
-                    throw new \Exception( curl_error($ch) );
+                    if (curl_errno($ch) == 6) {
+                        // Could not resolve host
+                        throw new \Exception('retry');
+                    }
+                    else {
+                        throw new \Exception( curl_error($ch) );
+                    }
                 }
 
                 // Do things with the response returned by the controller?
@@ -123,11 +127,24 @@ class MigrateCommand extends ContainerAwareCommand
 
             }
             catch (\Exception $e) {
-$output->writeln($e->getMessage());
-                $logger->err('MigrateCommand.php: '.$e->getMessage());
+                if ( $e->getMessage() == 'retry' ) {
+                    $output->writeln( 'Could not resolve host, releasing job to try again' );
+                    $logger->err('MigrateCommand.php: '.$e->getMessage());
 
-                // Delete the job so the queue hopefully doesn't hang
-                $pheanstalk->delete($job);
+                    // Release the job back into the ready queue to try again
+                    $pheanstalk->release($job);
+
+                    // Sleep for a bit
+                    usleep(1000000);     // sleep for 1 second
+                }
+                else {
+                    $output->writeln($e->getMessage());
+
+                    $logger->err('MigrateCommand.php: '.$e->getMessage());
+
+                    // Delete the job so the queue doesn't hang, in theory
+                    $pheanstalk->delete($job);
+                }
             }
         }
     }
