+++ src/ODR/AdminBundle/Command/XMLImportFileDownloadCommand.php
@@ -2,15 +2,15 @@
 
 /**
 * Open Data Repository Data Publisher
-* ImportValidate Command
+* XMLImportFileDownload Command
 * (C) 2015 by Nathan Stone (nate.stone@opendatarepository.org)
 * (C) 2015 by Alex Pires (ajpires@email.arizona.edu)
 * Released under the GPLv2
 *
 * This Symfony console command takes beanstalk jobs from the
-* validate_import tube and passes the parameters to XMLController,
-* which will attempt to validate the XML file against an XSD schema
-* prior to actually attempting to import it.
+* import_file tube and passes the parameters to XMLController,
+* which will attempt to download the described File or Image
+* entity from some remote server as part of an XML import.
 *
 */
 
@@ -29,15 +29,15 @@ use Symfony\Component\Console\Output\OutputInterface;
 use ODR\AdminBundle\Entity\DataRecord;
 use ODR\AdminBundle\Entity\DataType;
 
-class ImportValidateCommand extends ContainerAwareCommand
+class XMLImportFileDownloadCommand extends ContainerAwareCommand
 {
     protected function configure()
     {
         parent::configure();
 
         $this
-            ->setName('odr_import:validate')
-            ->setDescription('Attempts to validate an XML file scheduled for import...');
+            ->setName('odr_xml_import:file_download')
+            ->setDescription('Attempts to download a file from a remote server for importing purposes...');
     }
 
     protected function execute(InputInterface $input, OutputInterface $output)
@@ -45,33 +45,25 @@ class ImportValidateCommand extends ContainerAwareCommand
         // Only need to load these once...
         $container = $this->getContainer();
         $logger = $container->get('logger');
-        $router = $container->get('router');
         $pheanstalk = $container->get('pheanstalk');
 
         // Run command until manually stopped
         while (true) {
             $job = null;
             try {
-                // Wait for a job?
-//                $job = $pheanstalk->watch($memcached_prefix.'_validate_import')->ignore('default')->reserve();
-                $job = $pheanstalk->watch('validate_import')->ignore('default')->reserve();
+                // Wait for a job
+                $job = $pheanstalk->watch('import_file')->ignore('default')->reserve();
 
                 // Get Job Data
                 $data = json_decode($job->getData());
 
-                //
-                $str = '';
-                if ($data->datatype_id !== '')
-                    $str = 'Attempting to validate "'.$data->xml_filename.'" for DataType '.$data->datatype_id.' from '.$data->memcached_prefix.'...';
-                else if ($data->datafield_id !== '')
-                    $str = 'Attempting to validate "'.$data->xml_filename.'" for DataField '.$data->datafield_id.' from '.$data->memcached_prefix.'...';
-                else
-                    throw new \Exception('Invalid job data');
+                // 
+                $str = 'XMLImportFileDownloadCommand.php: Attempting to import the '.$data->object_type.' "'.$data->original_name.'" into drf '.$data->drf_id.' from '.$data->memcached_prefix.'...';
 
                 $current_time = new \DateTime();
-                $output->writeln( $current_time->format('Y-m-d H:i:s').' (UTC-5)' );                
+                $output->writeln( "\n".$current_time->format('Y-m-d H:i:s').' (UTC-5)' );
                 $output->writeln($str);
-                $logger->info('ImportValidateCommand.php: '.$str);
+                $logger->info('XMLImportFileDownloadCommand.php: '.$str);
 
                 // Need to use cURL to send a POST request...thanks symfony
                 $ch = curl_init();
@@ -79,7 +71,15 @@ class ImportValidateCommand extends ContainerAwareCommand
 $output->writeln($data->url);
 
                 // Create the required url and the parameters to send
-                $parameters = array('datatype_id' => $data->datatype_id, 'datafield_id' => $data->datafield_id, 'api_key' => $data->api_key, 'user_id' => $data->user_id, 'xml_filename' => $data->xml_filename);
+                $parameters = array(
+                    'object_type' => $data->object_type,
+                    'drf_id' => $data->drf_id,
+                    'user_id' => $data->user_id,
+                    'href' => $data->href,
+                    'original_name' => $data->original_name,
+                    'metadata' => $data->metadata,
+                    'api_key' => $data->api_key
+                );
 
                 // Set the options for the POST request
                 curl_setopt_array($ch, array(
@@ -96,22 +96,34 @@ $output->writeln($data->url);
 
                 // Send the request
                 if( ! $ret = curl_exec($ch)) {
-                    throw new \Exception( curl_error($ch) );
+                    if (curl_errno($ch) == 6) {
+                        // Could not resolve host
+                        throw new \Exception('retry');
+                    }
+                    else {
+                        throw new \Exception( curl_error($ch) );
+                    }
                 }
 
                 // Do things with the response returned by the controller?
                 $result = json_decode($ret);
                 if ( isset($result->r) && isset($result->d) ) {
-                    if ( $result->r == 0 )
+                    if ( $result->r == 0 ) {
                         $output->writeln( $result->d );
-                    else
+                    }
+                    else if ( $result->r == 6 ) {
+                        // Could not resolve host error
+                        throw new \Exception( 'retry' );
+                    }
+                    else {
                         throw new \Exception( $result->d );
+                    }
                 }
                 else {
                     // Should always be a json return...
                     throw new \Exception( print_r($ret, true) );
                 }
-//$logger->debug('ImportWorkerCommand.php: curl results...'.print_r($result, true));
+//$logger->debug('XMLImportFileDownloadCommand.php: curl results...'.print_r($result, true));
 
                 // Done with this cURL object
                 curl_close($ch);
@@ -121,11 +133,23 @@ $output->writeln($data->url);
 
             }
             catch (\Exception $e) {
-                $output->writeln($e->getMessage());
-                $logger->err('ImportValidateCommand.php: '.$e->getMessage());
+                if ( $e->getMessage() == 'retry' ) {
+                    $output->writeln( 'Could not resolve host, releasing job to try again' );
+                    $logger->err('XMLImportFileDownloadCommand.php: '.$e->getMessage());
 
-                // Delete the job so the queue doesn't hang, in theory
-                $pheanstalk->delete($job);
+                    // Release the job back into the ready queue to try again
+                    $pheanstalk->release($job);
+
+                    // Sleep for a bit
+                    usleep(1000000);     // sleep for 1 second
+                }
+                else {
+                    $output->writeln($e->getMessage());
+                    $logger->err('XMLImportFileDownloadCommand.php: '.$e->getMessage());
+
+                    // Delete the job so the queue doesn't hang, in theory
+                    $pheanstalk->delete($job);
+                }
             }
         }
     }
