+++ src/ODR/AdminBundle/Command/XMLImportStartCommand.php
@@ -2,7 +2,7 @@
 
 /**
 * Open Data Repository Data Publisher
-* ImportStart Command
+* XMLImportStart Command
 * (C) 2015 by Nathan Stone (nate.stone@opendatarepository.org)
 * (C) 2015 by Alex Pires (ajpires@email.arizona.edu)
 * Released under the GPLv2
@@ -30,14 +30,15 @@ use Symfony\Component\Console\Output\OutputInterface;
 use ODR\AdminBundle\Entity\DataRecord;
 use ODR\AdminBundle\Entity\DataType;
 
-class ImportStartCommand extends ContainerAwareCommand
+
+class XMLImportStartCommand extends ContainerAwareCommand
 {
     protected function configure()
     {
         parent::configure();
 
         $this
-            ->setName('odr_import:start')
+            ->setName('odr_xml_import:start')
             ->setDescription('Waits for an import request for a given datatype...');
     }
 
@@ -46,33 +47,25 @@ class ImportStartCommand extends ContainerAwareCommand
         // Only need to load these once...
         $container = $this->getContainer();
         $logger = $container->get('logger');
-        $router = $container->get('router');
         $pheanstalk = $container->get('pheanstalk');
 
         // Run command until manually stopped
         while (true) {
             $job = null;
             try {
-                // Wait for a job?
-//                $job = $pheanstalk->watch($memcached_prefix.'_import_datatype')->ignore('default')->reserve();
+                // Wait for a job
                 $job = $pheanstalk->watch('import_datatype')->ignore('default')->reserve();
 
                 // Get Job Data
                 $data = json_decode($job->getData());
 
                 // 
-                $str = '';
-                if ($data->datatype_id !== '')
-                    $str = 'Import request for DataType '.$data->datatype_id.' from '.$data->memcached_prefix.'...';
-                else if ($data->datafield_id !== '')
-                    $str = 'Import request for DataField '.$data->datafield_id.' from '.$data->memcached_prefix.'...';
-                else
-                    throw new \Exception('Invalid job data');
+                $str = 'XMLImport request for DataType '.$data->datatype_id.' from '.$data->memcached_prefix.'...';
 
                 $current_time = new \DateTime();
                 $output->writeln( $current_time->format('Y-m-d H:i:s').' (UTC-5)' );                
                 $output->writeln($str);
-                $logger->info('ImportStartCommand.php: '.$str);
+                $logger->info('XMLImportStartCommand.php: '.$str);
 
                 // Need to use cURL to send a POST request...thanks symfony
                 $ch = curl_init();
@@ -80,7 +73,7 @@ class ImportStartCommand extends ContainerAwareCommand
 $output->writeln($data->url);
 
                 // Create the required url and the parameters to send
-                $parameters = array('datatype_id' => $data->datatype_id, 'datafield_id' => $data->datafield_id, 'api_key' => $data->api_key, 'user_id' => $data->user_id);
+                $parameters = array('datatype_id' => $data->datatype_id, 'api_key' => $data->api_key, 'user_id' => $data->user_id);
 
                 // Set the options for the POST request
                 curl_setopt_array($ch, array(
@@ -97,7 +90,13 @@ $output->writeln($data->url);
 
                 // Send the request
                 if( ! $ret = curl_exec($ch)) {
-                    throw new \Exception( curl_error($ch) );
+                    if (curl_errno($ch) == 6) {
+                        // Could not resolve host
+                        throw new \Exception('retry');
+                    }
+                    else {
+                        throw new \Exception( curl_error($ch) );
+                    }
                 }
 
                 // Do things with the response returned by the controller?
@@ -112,21 +111,33 @@ $output->writeln($data->url);
                     // Should always be a json return...
                     throw new \Exception( print_r($ret, true) );
                 }
-//$logger->debug('ImportStartCommand.php: curl results...'.print_r($result, true));
+//$logger->debug('XMLImportStartCommand.php: curl results...'.print_r($result, true));
 
                 // Done with this cURL object
                 curl_close($ch);
 
                 // Dealt with (or ignored) the job
                 $pheanstalk->delete($job);
-
             }
             catch (\Exception $e) {
-                $output->writeln($e->getMessage());
-                $logger->err('ImportStartCommand.php: '.$e->getMessage());
+                if ( $e->getMessage() == 'retry' ) {
+                    $output->writeln( 'Could not resolve host, releasing job to try again' );
+                    $logger->err('XMLImportStartCommand.php: '.$e->getMessage());
 
-                // Delete the job so the queue doesn't hang, in theory
-                $pheanstalk->delete($job);
+                    // Release the job back into the ready queue to try again
+                    $pheanstalk->release($job);
+
+                    // Sleep for a bit
+                    usleep(1000000);     // sleep for 1 second
+                }
+                else {
+                    $output->writeln($e->getMessage());
+
+                    $logger->err('XMLImportStartCommand.php: '.$e->getMessage());
+
+                    // Delete the job so the queue doesn't hang, in theory
+                    $pheanstalk->delete($job);
+                }
             }
         }
     }
