+++ src/ODR/AdminBundle/Command/XMLImportWorkerCommand.php
@@ -2,14 +2,15 @@
 
 /**
 * Open Data Repository Data Publisher
-* Encrypt Command
+* XMLImportWorker Command
 * (C) 2015 by Nathan Stone (nate.stone@opendatarepository.org)
 * (C) 2015 by Alex Pires (ajpires@email.arizona.edu)
 * Released under the GPLv2
 *
 * This Symfony console command takes beanstalk jobs from the
-* mass_encrypt tube and passes the parameters to WorkerController
-* to either encrypt or decrypt a file or image entity.
+* import_worker tube and passes the parameters to XMLController,
+* which will attempt to import the contents of an XML file into
+* a DataRecord entity in the database.
 *
 */
 
@@ -28,51 +29,50 @@ use Symfony\Component\Console\Output\OutputInterface;
 use ODR\AdminBundle\Entity\DataRecord;
 use ODR\AdminBundle\Entity\DataType;
 
-class EncryptCommand extends ContainerAwareCommand
+
+class XMLImportWorkerCommand extends ContainerAwareCommand
 {
     protected function configure()
     {
         parent::configure();
 
         $this
-            ->setName('odr_encrypt:encrypt')
-            ->setDescription('Encrypts/Decrypts a File or Image object');
+            ->setName('odr_xml_import:worker')
+            ->setDescription('Attempts to import an XML file into the database...');
     }
 
     protected function execute(InputInterface $input, OutputInterface $output)
     {
         // Only need to load these once...
         $container = $this->getContainer();
-        $router = $container->get('router');
         $logger = $container->get('logger');
         $pheanstalk = $container->get('pheanstalk');
 
+        // Run command until manually stopped
         while (true) {
-            // Run command until manually stopped
             $job = null;
             try {
-                // Wait for a job?
-//                $job = $pheanstalk->watch($memcached_prefix.'_migrate_datafields')->ignore('default')->reserve(); 
-                $job = $pheanstalk->watch('mass_encrypt')->ignore('default')->reserve(); 
+                // Wait for a job
+                $job = $pheanstalk->watch('import_datarecord')->ignore('default')->reserve();
 
                 // Get Job Data
-                $data = json_decode($job->getData()); 
+                $data = json_decode($job->getData());
 
                 // 
-                $logger->info('EncryptCommand.php: Encrypt request for '.$data->object_type.' '.$data->object_id.' from '.$data->memcached_prefix.'...');
+                $str = 'XMLImportWorkerCommand.php: Attempting to import "'.$data->xml_filename.'" for DataType '.$data->datatype_id.' from '.$data->memcached_prefix.'...';
+
                 $current_time = new \DateTime();
-                $output->writeln( $current_time->format('Y-m-d H:i:s').' (UTC-5)' );                
-                $output->writeln('Encrypt request for '.$data->object_type.' '.$data->object_id.' from '.$data->memcached_prefix.'...');
+                $output->writeln( $current_time->format('Y-m-d H:i:s').' (UTC-5)' );
+                $output->writeln($str);
+                $logger->info('XMLImportWorkerCommand.php: '.$str);
 
                 // Need to use cURL to send a POST request...thanks symfony
                 $ch = curl_init();
 
-                // Create the required parameters to send
-                $parameters = array(
-                    'object_type' => $data->object_type,
-                    'object_id' => $data->object_id,
-                    'api_key' => $data->api_key
-                );
+$output->writeln($data->url);
+
+                // Create the required url and the parameters to send
+                $parameters = array('datatype_id' => $data->datatype_id, 'api_key' => $data->api_key, 'user_id' => $data->user_id, 'xml_filename' => $data->xml_filename);
 
                 // Set the options for the POST request
                 curl_setopt_array($ch, array(
@@ -82,14 +82,20 @@ class EncryptCommand extends ContainerAwareCommand
                         CURLOPT_FRESH_CONNECT => 1,
                         CURLOPT_RETURNTRANSFER => 1,
                         CURLOPT_FORBID_REUSE => 1,
-                        CURLOPT_TIMEOUT => 120, 
+                        CURLOPT_TIMEOUT => 300,
                         CURLOPT_POSTFIELDS => http_build_query($parameters)
                     )
                 );
 
                 // Send the request
                 if( ! $ret = curl_exec($ch)) {
-                    throw new \Exception( curl_error($ch) );
+                    if (curl_errno($ch) == 6) {
+                        // Could not resolve host
+                        throw new \Exception('retry');
+                    }
+                    else {
+                        throw new \Exception( curl_error($ch) );
+                    }
                 }
 
                 // Do things with the response returned by the controller?
@@ -104,25 +110,36 @@ class EncryptCommand extends ContainerAwareCommand
                     // Should always be a json return...
                     throw new \Exception( print_r($ret, true) );
                 }
-//$logger->debug('MigrateCommand.php: curl results...'.print_r($result, true));
+//$logger->debug('XMLImportWorkerCommand.php: curl results...'.print_r($result, true));
 
                 // Done with this cURL object
                 curl_close($ch);
 
-                // Dealt with the job
+                // Dealt with (or ignored) the job
                 $pheanstalk->delete($job);
-
-                // Sleep for a bit
-                usleep(200000);
 
             }
             catch (\Exception $e) {
-$output->writeln($e->getMessage());
-                $logger->err('EncryptCommand.php: '.$e->getMessage());
+                if ( $e->getMessage() == 'retry' ) {
+                    $output->writeln( 'Could not resolve host, releasing job to try again' );
+                    $logger->err('XMLImportWorkerCommand.php: '.$e->getMessage());
 
-                // Delete the job so the queue hopefully doesn't hang
-                $pheanstalk->delete($job);
+                    // Release the job back into the ready queue to try again
+                    $pheanstalk->release($job);
+
+                    // Sleep for a bit
+                    usleep(1000000);     // sleep for 1 second
+                }
+                else {
+                    $output->writeln($e->getMessage());
+
+                    $logger->err('XMLImportWorkerCommand.php: '.$e->getMessage());
+
+                    // Delete the job so the queue doesn't hang, in theory
+                    $pheanstalk->delete($job);
+                }
             }
         }
     }
+
 }
